requirements.txt
crewai>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.0.0
PyYAML>=6.0.1
requests>=2.31.0
beautifulsoup4>=4.12.0
openai>=1.12.0
langchain>=0.1.0
python-dateutil>=2.8.2
markdown2>=2.4.10 